requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.2.0
django-extensions==3.2.1
djangorestframework==3.14.0
dnspython==2.3.0
drf-extra-fields==3.7.0
filelock==3.10.7
filetype==1.2.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
fonttools==4.39.4
greenlet==2.0.2
idna==3.4
install==1.3.5
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
Libreria==0.1.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mysqlclient==2.2.0
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
platformdirs==3.2.0
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.7
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.21.0
Werkzeug==2.2.3
